refactor app.py and 'routes' for flask
